fixing object permissions checks

+++ lobbies/views.py
@@ -36,6 +36,7 @@ class LobbyView(viewsets.ModelViewSet):
     pk = request.data['lobby_id']
     queryset = Lobby.objects.all()
     lobby = get_object_or_404(queryset, pk=pk)
+    self.check_object_permissions(request, lobby)
     user = self.request.user
     if lobby.players.contains(user):
       return Response(data={
@@ -59,6 +60,7 @@ class LobbyView(viewsets.ModelViewSet):
     pk = request.data['lobby_id']
     queryset = Lobby.objects.all()
     lobby = get_object_or_404(queryset, pk=pk)
+    self.check_object_permissions(request, lobby)
     user = self.request.user
     if lobby.owner == user:
       lobby.delete()
@@ -78,6 +80,7 @@ class LobbyView(viewsets.ModelViewSet):
     pk = request.data['lobby_id']
     queryset = Lobby.objects.all()
     lobby = get_object_or_404(queryset, pk=pk)
+    self.check_object_permissions(request, lobby)
     if lobby.players.count() != 2:
       return Response(data={
           'detail': 'Lobby must be full, 2 player needed for this game.'
